% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/rowcol_macros.R
\name{numcols}
\alias{numcols}
\title{Return number of columns of dataframe
(this is slightly easier to read than dim(DF)[2])}
\usage{
numcols(DF = stop("DF not supplied"))
}
\arguments{
\item{DF}{a dataframe}
}
\value{
the number of columns of the dataframe
}
\description{
Return number of columns of dataframe
(this is slightly easier to read than dim(DF)[2])
}
\author{
Ben Veal
}
\seealso{
\code{\link{numrows}}
}

