% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/rowcol_macros.R
\name{numrows}
\alias{numrows}
\title{Return number of rows of dataframe
(this is slightly easier to read than dim(DF)[1])}
\usage{
numrows(DF = stop("DF not supplied"))
}
\arguments{
\item{DF}{a dataframe}
}
\value{
the number of rows of the dataframe
}
\description{
Return number of rows of dataframe
(this is slightly easier to read than dim(DF)[1])
}
\author{
Ben Veal
}
\seealso{
\code{\link{numcols}}
}

