% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/rowcol_macros.R
\name{rmgrep}
\alias{rmgrep}
\title{Remove from workspace all objects with names matching regexp}
\usage{
rmgrep(regexp = stop("regexp not supplied"))
}
\arguments{
\item{regexp}{a regular expression (as a character object)}
}
\value{
none - side effects only
}
\description{
Remove from workspace all objects with names matching regexp
}
\author{
Ben Veal
}
\seealso{
\code{\link{rmexgrep}}
}

