\name{strRmEmpty}
\alias{strRmEmpty}
\title{Macro to remove empty strings from character vector.}
\usage{
  strRmEmpty(strvec = stop("strvec not supplied"))
}
\arguments{
  \item{strvec}{a character vector}
}
\value{
  A character vector with empty strings removed
}
\description{
  Macro to remove empty strings from character vector.
}
\author{
  Ben Veal
}

