% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/rowcol_macros.R
\name{strRmSpace}
\alias{strRmSpace}
\title{Macro to remove whitespace from strings (inside string aswell as at ends).}
\usage{
strRmSpace(str = stop("str not supplied"))
}
\arguments{
\item{strvec}{a character vector}
}
\value{
A character vector with empty strings removed
}
\description{
Macro to remove whitespace from strings (inside string aswell as at ends).
}
\author{
Ben Veal
}

