% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/rowcol_macros.R
\name{strRmEmpty}
\alias{strRmEmpty}
\title{Macro to remove empty strings from character vector.}
\usage{
strRmEmpty(strvec = stop("strvec not supplied"))
}
\arguments{
\item{strvec}{a character vector}
}
\value{
A character vector with empty strings removed
}
\description{
Macro to remove empty strings from character vector.
}
\author{
Ben Veal
}

