\name{strRmSpace}
\alias{strRmSpace}
\title{Macro to remove whitespace from strings (inside string aswell as at ends).}
\usage{
  strRmSpace(str = stop("str not supplied"))
}
\arguments{
  \item{strvec}{a character vector}
}
\value{
  A character vector with empty strings removed
}
\description{
  Macro to remove whitespace from strings (inside string
  aswell as at ends).
}
\author{
  Ben Veal
}

